refactor mock disability rating service to separate class
